+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/resolver/LevelResolver.java
@@ -2,17 +2,39 @@ package com.vlkan.log4j2.logstash.layout.resolver;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import org.apache.logging.log4j.core.LogEvent;
+import org.apache.logging.log4j.core.net.Severity;
 
 import java.io.IOException;
 
 class LevelResolver implements EventResolver {
 
-    private static final LevelResolver INSTANCE = new LevelResolver();
+    private static final EventResolver NAME_RESOLVER = (logEvent, jsonGenerator) -> {
+        String levelName = logEvent.getLevel().name();
+        jsonGenerator.writeString(levelName);
+    };
 
-    private LevelResolver() {}
+    private static final EventResolver SEVERITY_NAME_RESOLVER = (logEvent, jsonGenerator) -> {
+        String severityName = Severity.getSeverity(logEvent.getLevel()).name();
+        jsonGenerator.writeString(severityName);
+    };
 
-    static LevelResolver getInstance() {
-        return INSTANCE;
+    private static final EventResolver SEVERITY_CODE_RESOLVER = (logEvent, jsonGenerator) -> {
+        int severityCode = Severity.getSeverity(logEvent.getLevel()).getCode();
+        jsonGenerator.writeNumber(severityCode);
+    };
+
+    private final EventResolver internalResolver;
+
+    LevelResolver(String key) {
+        if (key == null) {
+            internalResolver = NAME_RESOLVER;
+        } else if ("severity".equals(key)) {
+            internalResolver = SEVERITY_NAME_RESOLVER;
+        } else if ("severity:code".equals(key)) {
+            internalResolver = SEVERITY_CODE_RESOLVER;
+        } else {
+            throw new IllegalArgumentException("unknown key: " + key);
+        }
     }
 
     static String getName() {
@@ -21,8 +43,7 @@ class LevelResolver implements EventResolver {
 
     @Override
     public void resolve(LogEvent logEvent, JsonGenerator jsonGenerator) throws IOException {
-        String level = logEvent.getLevel().name();
-        jsonGenerator.writeString(level);
+        internalResolver.resolve(logEvent, jsonGenerator);
     }
 
 }